+++ integration-auth-commons/src/main/java/org/symphonyoss/integration/authentication/AuthenticationToken.java
