+++ integration-auth-commons/src/main/java/org/symphonyoss/integration/authentication/exception/UnexpectedAuthException.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright 2016-2017 Symphony Integrations - Symphony LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.symphonyoss.integration.authentication.exception;
+
+/**
+ * Failed to process certificate login.
+ *
+ * Created by ecarrenho on 26/10/16.
+ */
+public class UnexpectedAuthException extends AuthenticationException {
+
+  public UnexpectedAuthException(String message) {
+    super(message);
+  }
+
+  public UnexpectedAuthException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public UnexpectedAuthException(String message, Throwable cause, String... solutions) {
+    super(message, cause, solutions);
+  }
+}
