+++ integration-auth-commons/src/test/java/org/symphonyoss/integration/authentication/AuthenticationTokenTest.java
