+++ integration-logging/src/main/java/org/symphonyoss/integration/exception/ExceptionMessageFormatter.java
