+++ integration-logging/src/main/java/org/symphonyoss/integration/exception/IntegrationRuntimeException.java
