+++ integration-logging/src/main/java/org/symphonyoss/integration/logging/LogMessageSource.java
