+++ integration-logging/src/test/resources/integration-logging-log-messages.properties
@@ -1,0 +1,7 @@
+integration.commons.unable.to.load.log.messages = Unable to load log messages from log-messages.properties files. \
+  Integration Bridge logging will not display proper content.
+integration.commons.unable.to.load.log.messages.solution = This is probably related to a problem on the jar files \
+  for this Integration Bridge installation. Try restarting the Integration Bridge, and if that does not work, try \
+  reinstalling it.
+
+
